move editRole into own hook
