+++ src/components/pages/Roles/forms/RoleFormTabs.tsx
@@ -1,23 +1,17 @@
 import { Box, Tab, TabList, TabPanels, TabPanel, Tabs, Button, Flex } from '@chakra-ui/react';
 import { useFormikContext } from 'formik';
-import { useMemo, useEffect } from 'react';
+import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
-import { Hex, zeroAddress } from 'viem';
+import { Hex } from 'viem';
 import { DAO_ROUTES } from '../../../../constants/routes';
 import { useFractal } from '../../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../../providers/NetworkConfig/NetworkConfigProvider';
 import { useRolesState } from '../../../../state/useRolesState';
-import { EditBadgeStatus, EditedRole, RoleFormValues, RoleValue } from '../types';
+import { RoleFormValues } from '../types';
 import RoleFormInfo from './RoleFormInfo';
 import RoleFormPaymentStream from './RoleFormPaymentStream';
-
-const addRemoveField = (fieldNames: string[], fieldName: string, isRemoved: boolean) => {
-  if (fieldNames.includes(fieldName) && isRemoved) {
-    return fieldNames.filter(field => field !== fieldName);
-  }
-  return [...fieldNames, fieldName];
-};
+import { useRoleFormEditedRole } from './useRoleFormEditedRole';
 
 export default function RoleFormTabs({ hatId, push }: { hatId: Hex; push: (obj: any) => void }) {
   const { hatsTree } = useRolesState();
@@ -26,18 +20,9 @@ export default function RoleFormTabs({ hatId, push }: { hatId: Hex; push: (obj: 
   } = useFractal();
   const navigate = useNavigate();
   const { addressPrefix } = useNetworkConfig();
-
+  const { editedRoleData } = useRoleFormEditedRole({ hatsTree });
   const { t } = useTranslation(['roles']);
   const { values, errors, setFieldValue } = useFormikContext<RoleFormValues>();
-
-  const existingRoleHat = useMemo(
-    () =>
-      hatsTree?.roleHats.find(
-        (role: RoleValue) =>
-          !!values.roleEditing && role.id === values.roleEditing.id && role.id !== zeroAddress,
-      ),
-    [values.roleEditing, hatsTree],
-  );
 
   useEffect(() => {
     if (values.hats.length && !values.roleEditing) {
@@ -47,32 +32,6 @@ export default function RoleFormTabs({ hatId, push }: { hatId: Hex; push: (obj: 
       }
     }
   }, [values.hats, values.roleEditing, hatId, setFieldValue]);
-
-  const isRoleNameUpdated = !!existingRoleHat && values.roleEditing?.name !== existingRoleHat.name;
-
-  const isRoleDescriptionUpdated =
-    !!existingRoleHat && values.roleEditing?.description !== existingRoleHat.description;
-
-  const isMemberUpdated =
-    !!existingRoleHat && values.roleEditing?.wearer !== existingRoleHat.wearer;
-
-  const editedRole = useMemo<EditedRole>(() => {
-    if (!existingRoleHat) {
-      return {
-        fieldNames: [],
-        status: EditBadgeStatus.New,
-      };
-    }
-    let fieldNames: string[] = [];
-    fieldNames = addRemoveField(fieldNames, 'roleName', isRoleNameUpdated);
-    fieldNames = addRemoveField(fieldNames, 'roleDescription', isRoleDescriptionUpdated);
-    fieldNames = addRemoveField(fieldNames, 'member', isMemberUpdated);
-
-    return {
-      fieldNames,
-      status: EditBadgeStatus.Updated,
-    };
-  }, [existingRoleHat, isRoleNameUpdated, isRoleDescriptionUpdated, isMemberUpdated]);
 
   if (!daoAddress) return null;
 
@@ -99,7 +58,7 @@ export default function RoleFormTabs({ hatId, push }: { hatId: Hex; push: (obj: 
         <Button
           isDisabled={!!errors.roleEditing}
           onClick={() => {
-            const roleUpdated = { ...values.roleEditing, editedRole };
+            const roleUpdated = { ...values.roleEditing, editedRole: editedRoleData };
             const hatIndex = values.hats.findIndex(h => h.id === hatId);
             if (hatIndex === -1) {
               // @dev new hat
