+++ src/components/pages/Roles/forms/useRoleFormEditedRole.tsx
@@ -1,0 +1,53 @@
+import { useFormikContext } from 'formik';
+import { useMemo } from 'react';
+import { zeroAddress } from 'viem';
+import { DecentTree } from '../../../../state/useRolesState';
+import { RoleValue, EditedRole, EditBadgeStatus, RoleFormValues } from '../types';
+
+const addRemoveField = (fieldNames: string[], fieldName: string, isRemoved: boolean) => {
+  if (fieldNames.includes(fieldName) && isRemoved) {
+    return fieldNames.filter(field => field !== fieldName);
+  }
+  return [...fieldNames, fieldName];
+};
+
+export function useRoleFormEditedRole({ hatsTree }: { hatsTree: DecentTree | undefined | null }) {
+  const { values } = useFormikContext<RoleFormValues>();
+  const existingRoleHat = useMemo(
+    () =>
+      hatsTree?.roleHats.find(
+        (role: RoleValue) =>
+          !!values.roleEditing && role.id === values.roleEditing.id && role.id !== zeroAddress,
+      ),
+    [values.roleEditing, hatsTree],
+  );
+  const isRoleNameUpdated = !!existingRoleHat && values.roleEditing?.name !== existingRoleHat.name;
+
+  const isRoleDescriptionUpdated =
+    !!existingRoleHat && values.roleEditing?.description !== existingRoleHat.description;
+
+  const isMemberUpdated =
+    !!existingRoleHat && values.roleEditing?.wearer !== existingRoleHat.wearer;
+
+  const editedRoleData = useMemo<EditedRole>(() => {
+    if (!existingRoleHat) {
+      return {
+        fieldNames: [],
+        status: EditBadgeStatus.New,
+      };
+    }
+    let fieldNames: string[] = [];
+    fieldNames = addRemoveField(fieldNames, 'roleName', isRoleNameUpdated);
+    fieldNames = addRemoveField(fieldNames, 'roleDescription', isRoleDescriptionUpdated);
+    fieldNames = addRemoveField(fieldNames, 'member', isMemberUpdated);
+
+    return {
+      fieldNames,
+      status: EditBadgeStatus.Updated,
+    };
+  }, [existingRoleHat, isRoleNameUpdated, isRoleDescriptionUpdated, isMemberUpdated]);
+
+  return {
+    editedRoleData,
+  };
+}
